ML_Backend/requirements.txt
fastapi
uvicorn
pydantic
pymongo
sentence-transformers
python-dotenv